Change protocol usize's to u32

+++ shared/src/protocol.rs
@@ -17,7 +17,7 @@ pub enum PacketType {
 pub struct C2SReadMemoryPacket {
     _type: PacketType,
     pub address: u64,
-    pub size: usize,
+    pub size: u64,
 }
 
 #[derive(Debug, PartialEq, DekuRead, DekuWrite)]
@@ -59,7 +59,7 @@ pub struct C2STargetPidPacket {
 #[deku(endian = "big")]
 pub struct S2CReadMemoryPacketResponse {
     _type: PacketType,
-    pub count: usize,
+    pub count: u32,
     #[deku(count = "count")]
     pub data: Vec<u8>,
 }
@@ -68,7 +68,7 @@ pub struct S2CReadMemoryPacketResponse {
 #[deku(endian = "big")]
 pub struct S2CWriteMemoryPacketResponse {
     _type: PacketType,
-    pub bytes_written: usize,
+    pub bytes_written: u32,
 }
 
 #[derive(Debug, PartialEq, DekuRead, DekuWrite)]
@@ -111,7 +111,7 @@ impl C2SReadMemoryPacket {
         value
     }
 
-    pub fn out_bytes(address: u64, size: usize) -> Vec<u8> {
+    pub fn out_bytes(address: u64, size: u64) -> Vec<u8> {
         let object = C2SReadMemoryPacket {
             _type: PacketType::Read,
             address,
@@ -147,7 +147,7 @@ impl S2CReadMemoryPacketResponse {
     pub fn out_bytes(data: Vec<u8>) -> Vec<u8> {
         let object = S2CReadMemoryPacketResponse {
             _type: PacketType::Read,
-            count: data.len(),
+            count: data.len() as u32,
             data,
         };
         object.to_bytes().unwrap()
@@ -159,7 +159,7 @@ impl S2CWriteMemoryPacketResponse {
         let (_, value) = S2CWriteMemoryPacketResponse::from_bytes((data, 0)).unwrap();
         value
     }
-    pub fn out_bytes(bytes_written: usize) -> Vec<u8> {
+    pub fn out_bytes(bytes_written: u32) -> Vec<u8> {
         let object = S2CWriteMemoryPacketResponse {
             _type: PacketType::Write,
             bytes_written,
@@ -262,7 +262,7 @@ mod tests {
 
     #[test]
     fn test_write_memory_packet_response() {
-        const BYTES_WRITTEN: usize = 100;
+        const BYTES_WRITTEN: u32 = 100;
         let response_data = S2CWriteMemoryPacketResponse::out_bytes(BYTES_WRITTEN);
         let parsed_response = S2CWriteMemoryPacketResponse::parse(&response_data);
